refactor: 1차 파일 정리
