+++ src/refactoring/components/admin/coupon/index.ts
@@ -1,0 +1,4 @@
+import AddCouponForm from "./AddCouponForm";
+import Coupons from "./Coupons";
+
+export { AddCouponForm, Coupons };
