+++ src/refactoring/components/admin/product/index.ts
@@ -1,0 +1,4 @@
+import AddProductSection from "./AddProductSection";
+import Products from "./Products";
+
+export { AddProductSection, Products };
